Being able to register a company

+++ internals/models/models.go
@@ -15,7 +15,7 @@ type Company struct {
 	Base
 	Ruc       string `json:"ruc" gorm:"not null;index;unique"`
 	Name      string `json:"name" gorm:"not null;index;unique"`
-	Employees uint   `json:"employees" gorm:"not null;default:1;type:smallint"`
+	Employees uint8  `json:"employees" gorm:"not null;default:1;type:smallint"`
 	IsActive  bool   `json:"is_active" gorm:"not null;default:true"`
 }
